feat(web): decouple subagents into own module + first-class Agents dock tab
Move the subagent-lineage UI (SubagentSection/row tree, SessionAgentsChip,
SpawnedByBreadcrumb, sessionStatusTone) out of goal-surface.tsx into a new
session/subagents.tsx, so a session's agent tree is no longer coupled to
whether it carries a goal. Add an AgentsPanel and wire it as a first-class
"Agents" right-dock tab (with a live count badge) that appears whenever the
session has children, fed by a shared-feed useSessionLineage read. Remove the
Subagents section from the goal panel; repoint rail-shell imports.
